+++ bootique-cayenne/src/main/java/io/bootique/cayenne/ServerRuntimeFactory.java
@@ -30,11 +30,7 @@ import org.apache.cayenne.configuration.server.ServerRuntimeBuilder;
 import org.apache.cayenne.di.Key;
 import org.apache.cayenne.di.Module;
 
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.Collections;
-import java.util.LinkedHashSet;
-import java.util.List;
+import java.util.*;
 
 @BQConfig("Configures Cayenne stack, providing injectable ServerRuntime.")
 public class ServerRuntimeFactory {
@@ -43,13 +39,18 @@ public class ServerRuntimeFactory {
 
     private String name;
     private Collection<String> configs;
-    private List<DataMapConfig> maps;
+    /**
+     * Changed from list to map
+     *
+     * @since 2.0
+     */
+    private Map<String, DataMapConfig> maps;
     private String datasource;
     private boolean createSchema;
 
     public ServerRuntimeFactory() {
         this.configs = new ArrayList<>();
-        this.maps = new ArrayList<>();
+        this.maps = new HashMap<>();
     }
 
     public ServerRuntime createCayenneRuntime(
@@ -86,7 +87,8 @@ public class ServerRuntimeFactory {
 
             DefaultDataSourceName defaultDataSourceName = defaultDataSourceName(dataSourceFactory);
             binder.bind(Key.get(DefaultDataSourceName.class)).toInstance(defaultDataSourceName);
-            binder.bindList(DataMapConfig.class).addAll(maps);
+            binder.bindMap(DataMapConfig.class).putAll(maps);
+
 
             // provide default DataNode
             // TODO: copied from Cayenne, as the corresponding provider is not public or rather
@@ -152,14 +154,17 @@ public class ServerRuntimeFactory {
     }
 
     /**
-     * Sets a list of DataMaps that are included in the app runtime without an explicit refrence in  'cayenne-project.xml'.
+     * Sets a map of DataMaps that are included in the app runtime without an explicit refrence in  'cayenne-project.xml'.
      *
-     * @param maps list of DataMap configs
+     * Parameter maps changed from list type to map type
+     *
+     * @param maps map of DataMap configs
      * @since 0.18
+     * @since 2.0
      */
     @BQConfigProperty("A list of DataMaps that are included in the app runtime without an explicit refrence in " +
             "'cayenne-project.xml'.")
-    public void setMaps(List<DataMapConfig> maps) {
+    public void setMaps(Map<String, DataMapConfig> maps) {
         this.maps = maps;
     }